Refactored actor_critic_network into its own module directory
